refactored time_series unit tests to run faster in parallel
